Move embeddings into main investigation view
