+++ web/src/router/index.ts
@@ -2,7 +2,6 @@ import Vue from 'vue';
 import VueRouter, { RouteConfig } from 'vue-router';
 import InvestigationsList from '../views/InvestigationsList.vue';
 import InvestigationDetail from '../views/InvestigationDetail.vue';
-import EmbeddingsDetail from '../views/EmbeddingsDetail.vue';
 
 Vue.use(VueRouter);
 
@@ -16,12 +15,6 @@ const routes: Array<RouteConfig> = [
     path: '/investigations/:investigation',
     name: 'investigationDetail',
     component: InvestigationDetail,
-    props: true,
-  },
-  {
-    path: '/investigations/:investigation/embeddings',
-    name: 'embeddingsDetail',
-    component: EmbeddingsDetail,
     props: true,
   },
 ];
