+++ web/src/views/InvestigationDetail.vue
@@ -104,7 +104,13 @@
 
 <script lang="ts">
 import {
-  ref, defineComponent, onMounted, PropType, computed, watch, Ref,
+  ref,
+  defineComponent,
+  onMounted,
+  PropType,
+  computed,
+  watch,
+  Ref,
 } from '@vue/composition-api';
 import useGeoJS from '../utilities/useGeoJS';
 import { Point, postGisToPoint } from '../utilities/utiltyFunctions';
@@ -112,6 +118,26 @@ import store from '../store';
 import InvestigationSidebar from '../components/InvestigationSidebar.vue';
 import InvestigationDetailFrameMenu from '../components/InvestigationDetailFrameMenu.vue';
 import { Dataset, Pin } from '../generatedTypes/AtlascopeTypes';
+import type { DatasetEmbedding } from '../generatedTypes/AtlascopeTypes';
+
+interface RootDatasetEmbedding {
+  id: null;
+  child_bounding_box: number[];
+  parent: null;
+  child: string;
+}
+
+interface StackFrame {
+  embedding: DatasetEmbedding | RootDatasetEmbedding;
+  parent: {
+    scale: number;
+    offset: {
+      x: number;
+      y: number;
+    };
+  };
+  treeDepth: number;
+}
 
 export default defineComponent({
   name: 'InvestigationDetail',
@@ -190,26 +216,129 @@ export default defineComponent({
       if (!dataset || !dataset.id) {
         return;
       }
-      const tileSourceMetadata = store.state.datasetTileMetadata[dataset.id];
-      if (!tileSourceMetadata) {
-        return;
-      }
-      const geojsParams = generatePixelCoordinateParams(
-        tileSourceMetadata.size_x || 0,
-        tileSourceMetadata.size_y || 0,
-        tileSourceMetadata.tile_size || 0,
-        tileSourceMetadata.tile_size || 0,
-      );
-      if (!geojsParams || !geojsParams.map || !geojsParams.layer) {
-        return;
-      }
-      const apiRoot = process.env.VUE_APP_API_ROOT;
-      const queryString = buildUrlQueryArgs();
-      geojsParams.layer.url = `${apiRoot}/datasets/${dataset.id}/tiles/{z}/{x}/{y}.png${queryString}`;
-      geojsParams.layer.crossDomain = 'use-credentials';
 
-      createMap(geojsParams.map);
-      activeDatasetLayer.value = createLayer('osm', geojsParams.layer);
+      const apiRoot = process.env.VUE_APP_API_ROOT;
+      const embeddings = store.state.datasetEmbeddings;
+      const rootDatasetID = dataset.id;
+      const rootTileMetadata = store.state.datasetTileMetadata[rootDatasetID];
+
+      if (
+        rootTileMetadata === undefined
+        || rootTileMetadata.size_x === undefined
+        || rootTileMetadata.size_y === undefined
+        || rootTileMetadata.tile_size === undefined
+        || rootTileMetadata.levels === undefined
+      ) {
+        return;
+      }
+
+      const rootPixelParams = generatePixelCoordinateParams(
+        rootTileMetadata.size_x || 0,
+        rootTileMetadata.size_y || 0,
+        rootTileMetadata.tile_size || 0,
+        rootTileMetadata.tile_size || 0,
+      );
+      const mapParams = {
+        ...rootPixelParams.map,
+        max: 40,
+      };
+      const rootLayerParams = {
+        ...rootPixelParams.layer,
+        zIndex: 0,
+        url: `${apiRoot}/datasets/${rootDatasetID}/tiles/{z}/{x}/{y}.png`,
+        crossDomain: 'use-credentials',
+      };
+      createMap(mapParams);
+      rootDatasetLayer.value = createLayer('osm', rootLayerParams);
+
+      const visited: Set<RootDatasetEmbedding | DatasetEmbedding> = new Set();
+      const stack: Array<StackFrame> = [];
+      stack.unshift(
+        ...embeddings
+          .filter((e) => e.parent === rootDatasetID)
+          .map((e) => ({
+            embedding: e,
+            parent: { scale: 1, offset: { x: 0, y: 0 } },
+            treeDepth: 1,
+          })),
+      );
+
+      while (stack.length > 0) {
+        const { embedding, parent, treeDepth } = stack.shift()!;
+        if (!visited.has(embedding)) {
+          visited.add(embedding);
+
+          if (
+            embedding === undefined
+            || embedding.child_bounding_box === undefined
+            || embedding.child_bounding_box.length !== 4
+          ) {
+            return;
+          }
+
+          const datasetID = embedding.child;
+          const tileMetadata = store.state.datasetTileMetadata[datasetID];
+
+          if (
+            tileMetadata === undefined
+            || tileMetadata.size_x === undefined
+            || tileMetadata.size_y === undefined
+            || tileMetadata.tile_size === undefined
+            || tileMetadata.levels === undefined
+          ) {
+            return;
+          }
+
+          const boundingBox = {
+            x: {
+              min: embedding.child_bounding_box[0],
+              max: embedding.child_bounding_box[2],
+            },
+            y: {
+              min: embedding.child_bounding_box[1],
+              max: embedding.child_bounding_box[3],
+            },
+          };
+          const scale = Math.min(
+            (parent.scale * (boundingBox.x.max - boundingBox.x.min))
+              / tileMetadata.size_x,
+            (parent.scale * (boundingBox.y.max - boundingBox.y.min))
+              / tileMetadata.size_y,
+          );
+          const offset = {
+            x: (parent.scale / scale) * (parent.offset.x + boundingBox.x.min),
+            y: (parent.scale / scale) * (parent.offset.y + boundingBox.y.min),
+          };
+          const pixelParams = generatePixelCoordinateParams(
+            tileMetadata.size_x || 0,
+            tileMetadata.size_y || 0,
+            tileMetadata.tile_size || 0,
+            tileMetadata.tile_size || 0,
+          );
+          const layerParams = {
+            ...pixelParams.layer,
+            zIndex: treeDepth,
+            url: `${apiRoot}/datasets/${datasetID}/tiles/{z}/{x}/{y}.png`,
+            crossDomain: 'use-credentials',
+          };
+          createLayer(
+            'osm',
+            layerParams,
+            `+proj=longlat +axis=enu +xoff=-${offset.x} +yoff=${
+              offset.y
+            } +s11=${1 / scale} +s22=${1 / scale}`,
+          );
+
+          const frontier = embeddings.filter((e) => e.parent === datasetID);
+          stack.unshift(
+            ...frontier.map((e) => ({
+              embedding: e,
+              parent: { scale, offset },
+              treeDepth: treeDepth + 1,
+            })),
+          );
+        }
+      }
       clampBoundsX(false);
     }
