+++ web/src/views/InvestigationsList.vue
@@ -38,14 +38,6 @@
                 Investigate
               </v-btn>
             </router-link>
-            <router-link :to="`/investigations/${investigation.id}/embeddings`">
-              <v-btn
-                color="blue"
-                text
-              >
-                Embeddings
-              </v-btn>
-            </router-link>
           </v-card-actions>
         </v-card>
       </v-col>
